Book-Flight Form design
design finished
todos:
- object as prop
- spinner and date picker
- upload object to user collection with firestore
