+++ FlightBookingApp/App.js
@@ -1,47 +1,46 @@
 import * as React from 'react';
 import {NavigationContainer} from '@react-navigation/native';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
-import {SignUp, Login, My_flights, Origen} from './src/Screens/index';
+import {SignUp, Login, MyFlights, Origen, Destino, Fecha, Pasajeros, Confirmacion} from './src/Screens/index';
 
 const Stack = createNativeStackNavigator();
 
 const App = () => {
+
+  const stackOptions = {
+    headerStyle: {
+      elevation: 0,
+      shadowOpacity: 0,
+    },
+    headerTintColor: '#5974f5', 
+    headerTitleStyle: {
+      fontSize: 30, 
+      fontWeight: 'bold',
+      left: 0,
+    }, 
+    headerTitleAlign:'left',
+  }
+  const stackOptionsForm = {
+    headerStyle: {
+      elevation: 0,
+      shadowOpacity: 0,
+    },
+    headerTintColor: '#5974f5', 
+    title: '',
+  }
+
   return (
     <NavigationContainer>
       {
         <Stack.Navigator>
-          <Stack.Screen name="Login" component={Login}
-            options={{
-              headerTintColor: '#5974f5',
-              headerTitleStyle: {
-                fontWeight: 'bold',
-              },
-            }}
-          />
-          <Stack.Screen name="SignUp" component={SignUp}
-            options={{
-              headerTintColor: '#5974f5',
-              headerTitleStyle: {
-                fontWeight: 'bold',
-              },
-            }}
-          />
-          {/* <Stack.Screen name="My Flights" component={My_flights}
-            options={{
-              headerTintColor: '#5974f5',
-              headerTitleStyle: {
-                fontWeight: 'bold',
-              },
-            }}
-          /> */}
-          <Stack.Screen name="Origen" component={Origen}
-            options={{
-              headerTintColor: '#5974f5',
-              headerTitleStyle: {
-                fontWeight: 'bold',
-              },
-            }}
-          />
+          <Stack.Screen name="Login" component={Login} options={{stackOptions}}/>
+          <Stack.Screen name="SignUp" component={SignUp} options={{stackOptions}}/>
+          <Stack.Screen name="My Flights" component={MyFlights} options={{stackOptions}}/>
+          <Stack.Screen name="Origen" component={Origen} options={{stackOptionsForm}}/>
+          <Stack.Screen name='Destino' component={Destino} options={stackOptionsForm}/>
+          <Stack.Screen name='Fecha' component={Fecha} options={stackOptionsForm}/>
+          <Stack.Screen name='Pasajeros' component={Pasajeros} options={stackOptionsForm}/>
+          <Stack.Screen name='Confirmacion' component={Confirmacion} options={stackOptionsForm}/>
         </Stack.Navigator>
       }
     </NavigationContainer>
