+++ FlightBookingApp/src/Screens/MyFlights.js
@@ -2,7 +2,6 @@ import React, { useState, useEffect } from 'react';
 import { Text, View, TouchableOpacity } from 'react-native';
 import {Container2} from '../Assets/cardStyled'
 import { PlussButton, MyFlights_List } from '../Components/CardFlight/index';
-//import db from '../firebase/config';
 
 export const MyFlights = (props) => {
