+++ FlightBookingApp/src/Screens/index.js
@@ -2,3 +2,8 @@ export {SignUp} from "./SignUp";
 export {Login} from "./Login";
 export {MyFlights} from "./MyFlights";
 export {AddFlight} from "./AddFlight";
+export {Origen} from '../Components/MyFlightComponents/Origen'
+export {Destino} from '../Components/MyFlightComponents/Destino'
+export {Fecha} from '../Components/MyFlightComponents/Fecha'
+export {Pasajeros} from '../Components/MyFlightComponents/Pasajeros'
+export {Confirmacion} from '../Components/MyFlightComponents/Confirmacion'
